moved jump test to AbstractShapeTest
